requirements.txt
Django==4.2.0
psycopg2-binary==2.9.6
django-environ==0.11.2
Faker==0.7.4
django-rest-framework==0.1.0